move files and update paths
