+++ frontend/pickquick/src/App.jsx
@@ -5,10 +5,10 @@ import FilterContainer from "./components/FilterContainer";
 import Home from "./components/Home";
 import About from "./components/About";
 import Footer from "./components/Footer";
-import Selection from "./components/Selection";
-import Recommendations from "./components/Recommendations";
+import Selection from "./components/movie-components/Selection";
+import Recommendations from "./components/movie-components/Recommendations";
 import Feedback from "./components/Feedback";
-import MovieCard from "./components/MovieCard";
+import MovieCard from "./components/movie-components/MovieCard";
 import UserDashboard from "./components/gated/UserDashboard";
 import { BackendContext } from "./context/UserSyncHandler";
 import { useContext } from "react";
