+++ frontend/pickquick/src/components/movie-components/Recommendations.jsx
@@ -1,6 +1,6 @@
-import { useAppContext } from "../context/AppContext.jsx";
-import MovieCard from "./MovieCard";
-import "../styles/recommend-movieCard.css";
+import { useAppContext } from "../../context/AppContext.jsx";
+import MovieCard from "./MovieCard.jsx";
+import "../../styles/recommend-movieCard.css";
 
 const Recommendations = () => {
   const { recommendations } = useAppContext();
