+++ frontend/pickquick/src/components/movie-components/SelectionTrailersCard.jsx
